requirements.txt
streamlit
dotenv

azure.core
azure-ai-documentintelligence
streamlit
azure-storage-blob
